refactor: reduce GUI status adapter complexity
